requirements.txt
# Discord Webhook Manager
discord.py

# Music Download
yt-dlp


# Env
dotenv